% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_primary_key.R
\name{composite_primary_keys}
\alias{composite_primary_keys}
\title{Find what columns a data set is unique by}
\usage{
composite_primary_keys(data, expected_primary_keys)
}
\arguments{
\item{data}{Dataframe to find primary keys of}

\item{expected_primary_keys}{List form of columns that should identify the dataset uniquely}
}
\description{
Find what columns a data set is unique by
}
\examples{
data <- data.frame(a = c(1, 2, 3, 4, 5, 6),
                        b = c(1, 2, 2, 3, 4, 5),
                        c = c('a', 'b', 'b', 'c', 'd','e'),
                        id = c(1111, 2222, 2222, 3333, 4444, 5555))
nrow(data)
n_distinct(data$id)
primary_keys <- composite_primary_keys(data, c("id"))
print(primary_keys)

composite_primary_keys()
}
